backend/pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.1.1",
    "hf-xet>=1.1.10",
    "ipython>=9.6.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.30",
    "langchain-google-genai>=2.1.12",
    "langchain-openai>=0.3.35",
    "pdf2image>=1.17.0",
    "python-dotenv>=1.1.1",
    "tiktoken>=0.12.0",
    "unstructured[all-docs]>=0.18.15",
]
